move manifest to its own file
- add a dedicated loader
- make sure that the version starts with a 'v'
